test lock: ok avec un fpga, hs avec l'autre...

+++ userspace/lock_interrupts.c
@@ -42,7 +42,7 @@ static bool send_async(message_t* message) {
 }
 
 static bool send_lock_data(off_t offset, size_t nbytes, int isFull) {
-	int16_t* buffer = malloc(nbytes);
+	int64_t* buffer = malloc(nbytes);
 	if (buffer == NULL) {
 		log_error_errno("Unable to malloc buffer of %d bytes", nbytes);
 		return false;
@@ -60,7 +60,7 @@ static bool send_lock_data(off_t offset, size_t nbytes, int isFull) {
 	clock_gettime(CLOCK_MONOTONIC, &tend);
 	log_info("read lock data (%d bytes): %.3f ms", nbytes,
 		(tend.tv_sec - tstart.tv_sec) * 1000 + (tend.tv_nsec - tstart.tv_nsec) / 1000000.0f);
-
+	
 	message_t* message = create_message_with_body(MSG_LOCK_SCAN_DONE, buffer, nbytes);
 	if (message == NULL) {
 		free(buffer);
